rename ClaimEditor -> EntityValueInput
