+++ app/modules/inventory/components/entity_value_input.svelte
